+++ src/main/java/com/example/repository/DeptRepository.java
@@ -1,9 +1,11 @@
 package com.example.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 
 import com.example.domain.Dept;
 
-public interface DeptRepository extends CrudRepository<Dept, Integer> {
+public interface DeptRepository extends CrudRepository<Dept, BigInteger> {
 
 }
